demo/requirements.txt
attrs==21.4.0
Authlib==1.0.0
bootstrapped==0.0.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.2
cycler==0.11.0
dnspython==2.2.1
email-validator==1.1.3
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Dance==5.1.0
flask-mongoengine==1.0.0
Flask-WTF==1.0.1
fonttools==4.31.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
kiwisolver==1.4.1
MarkupSafe==2.1.1
matplotlib==3.5.1
mongoengine==0.24.1
mpmath==1.2.1
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.1
parameterized==0.8.1
patsy==0.5.2
Pillow==9.0.1
pipfile==0.0.2
pycparser==2.21
pymongo==4.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
requestes==0.0.1
requests==2.27.1
requests-oauthlib==1.3.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
statsmodels==0.13.2
sympy==1.10.1
tealang==0.4.0
threadpoolctl==3.1.0
toml==0.10.2
urllib3==1.26.9
URLObject==2.4.3
Werkzeug==2.0.3
WTForms==3.0.1
z3-solver==4.8.15.0
